Added Prefix for organization.
Prefix is created for organizing Prefix stuff. Prefix gets stored as a HashMap.

Added `/sp reload` to reload the plugin configuration.

Invalid Prefix IDs will now fallback to the default instead of getting angry and throwing an exception.

verify-always setting now verifies the prefix anytime PlayerUtil.getPlayerPrefix() is called. If the prefix is not earned, the prefix saved is cleared and the default prefix is used as a fallback.

Started work on PrefixMenu for Chest GUIs but it is non-functional. Compilation errors removed so the plugin can compile for other testing.