Testing Twitch HTTP Connections
Worked on this for a while, but could not commit earlier cause API Keys and what not.

Seems to work, need to perform other GET requests in order to obtain streaming information as well as subscribe to a stream when a player logs in and unsubscribe when a player logs out.